Blacklist

+++ sql/ddl-update21.sql
@@ -1,1 +1,1 @@
-alter table generic_log add blacklist_level integer null after remote_host;
+alter table generic_log add blacklist_level numeric(5,0) null after remote_host;
